Code refactor, add new modules
