+++ src/manifests/catalogs.rs
@@ -1,11 +1,12 @@
 use crate::api::schema::*;
 use crate::log::logging::*;
+
 use std::error::Error;
 use std::fs;
 use std::fs::File;
 use std::io::Read;
 
-// read_operator_catalog - simple function tha treads the specific catalog.json file
+// read_operator_catalog - simple function that reads the specific catalog.json file
 // and unmarshals it to DeclarativeConfig struct
 pub fn read_operator_catalog(path: String) -> Result<serde_json::Value, Box<dyn Error>> {
     let catalog = path + &"/catalog.json".to_owned();
@@ -32,7 +33,6 @@ pub fn read_operator_catalog(path: String) -> Result<serde_json::Value, Box<dyn 
 
 // find a specifc directory in the untar layers
 pub async fn find_dir(log: &Logging, dir: String, name: String) -> String {
-    // env::set_current_dir("../../../../").expect("could not set current directory");
     let paths = fs::read_dir(&dir);
     // for both release & operator image indexes
     // we know the layer we are looking for is only 1 level
@@ -60,4 +60,86 @@ pub async fn find_dir(log: &Logging, dir: String, name: String) -> String {
         }
     }
     return "".to_string();
+}
+
+// parse the manifest json for operator indexes only
+pub fn parse_json_manifest(data: String) -> Result<ManifestSchema, Box<dyn std::error::Error>> {
+    // Parse the string of data into serde_json::ManifestSchema.
+    let root: ManifestSchema = serde_json::from_str(&data)?;
+    Ok(root)
+}
+
+// parse the manifest json for operators
+pub fn parse_json_manifest_operator(data: String) -> Result<Manifest, Box<dyn std::error::Error>> {
+    // Parse the string of data into serde_json::Manifest.
+    let root: Manifest = serde_json::from_str(&data)?;
+    Ok(root)
+}
+
+// parse the manifest list json
+pub fn parse_json_manifestlist(data: String) -> Result<ManifestList, Box<dyn std::error::Error>> {
+    // Parse the string of data into serde_json::ManifestList.
+    let root: ManifestList = serde_json::from_str(&data)?;
+    Ok(root)
+}
+
+// contruct the manifest url
+pub fn get_image_manifest_url(image_ref: ImageReference) -> String {
+    // return a string in the form of (example below)
+    // "https://registry.redhat.io/v2/redhat/certified-operator-index/manifests/v4.12";
+    let mut url = String::from("https://");
+    url.push_str(&image_ref.registry);
+    url.push_str(&"/v2/");
+    url.push_str(&image_ref.namespace);
+    url.push_str(&"/");
+    url.push_str(&image_ref.name);
+    url.push_str(&"/");
+    url.push_str(&"manifests/");
+    url.push_str(&image_ref.version);
+    url
+}
+
+// contruct a manifest url from a string
+pub fn get_manifest_url(url: String) -> String {
+    let mut parts = url.split("/");
+    let mut url = String::from("https://");
+    url.push_str(&parts.nth(0).unwrap());
+    url.push_str(&"/v2/");
+    url.push_str(&parts.nth(0).unwrap());
+    url.push_str(&"/");
+    let i = parts.nth(0).unwrap();
+    let mut sha = i.split("@");
+    url.push_str(&sha.nth(0).unwrap());
+    url.push_str(&"/");
+    url.push_str(&"manifests/");
+    url.push_str(&sha.nth(0).unwrap());
+    url
+}
+
+// contruct a manifest url from a string by digest
+pub fn get_manifest_url_by_digest(url: String, digest: String) -> String {
+    let mut parts = url.split("/");
+    let mut url = String::from("https://");
+    url.push_str(&parts.nth(0).unwrap());
+    url.push_str(&"/v2/");
+    url.push_str(&parts.nth(0).unwrap());
+    url.push_str(&"/");
+    let i = parts.nth(0).unwrap();
+    let mut sha = i.split("@");
+    url.push_str(&sha.nth(0).unwrap());
+    url.push_str(&"/");
+    url.push_str(&"manifests/");
+    url.push_str(&digest);
+    url
+}
+
+// utility functions - get_manifest_json
+pub fn get_manifest_json_file(name: String, version: String) -> String {
+    let mut file = String::from("working-dir/");
+    file.push_str(&name);
+    file.push_str(&"/");
+    file.push_str(&version);
+    file.push_str(&"/");
+    file.push_str(&"manifest.json");
+    file
 }
